Pset1
MARIO PYRAMID
Write, in a program that recreates this half-pyramid using hashes (#) for blocks. However, to make things more interesting, first prompt the user for the half-pyramid’s height, a non-negative integer no greater than 23. If the user fails to provide a non-negative integer no greater than 23, you should re-prompt for the same again. Then, generate the desired half-pyramid. Take care to align the bottom-left corner of your half-pyramid with the left-hand edge of your terminal window, as in the sample output below, wherein underlined text represents some user’s input.

GREEDY ALGORITHM
Write, in a program that first asks the user how much change is owed and then spits out the minimum number of coins with which said change can be made.